# Project/backend/swagger.yaml
openapi: 3.0.0
info:
    title: 'HBL Banking API'
    version: 1.0.0
    description: 'A comprehensive banking API for HBL Clone application'
    contact:
        name: 'Muhammad Junaid Afzal'
        email: me@junaidafzal.dev
    license:
        name: ISC
        url: 'https://opensource.org/licenses/ISC'
servers:
    -
        url: 'http://localhost:8000'
        description: 'Development server'
    -
        url: 'https://your-production-url.com'
        description: 'Production server'
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
...
        Report:
            type: object
            properties: {reportId: {type: string, example: RPT-1234567890-ABC123}, category: {type: string, enum: [account_issue, transaction_problem, card_issue, loan_problem, technical_issue, complaint, suggestion, other]}, subject: {type: string, maxLength: 200}, description: {type: string, maxLength: 2000}, priority: {type: string, enum: [low, medium, high, urgent]}, status: {type: string, enum: [submitted, in_progress, resolved, closed]}, images: {type: array, items: {type: object, properties: {url: {type: string}, publicId: {type: string}, uploadedAt: {type: string, format: date-time}}}}, submittedAt: {type: string, format: date-time}, lastUpdated: {type: string, format: date-time}, resolution: {type: object, properties: {resolvedAt: {type: string, format: date-time}, resolvedBy: {type: string}, resolutionNote: {type: string}}}}
...
security:
    -
        bearerAuth: []
paths:
...
    '/api/admin/reports/{type}':
        get:
            summary: 'Generate system reports (Admin only)'
            tags: [Admin]
            security: [{bearerAuth: []}]
            parameters: [{in: path, name: type, required: true, schema: {type: string, enum: [transactions, accounts, users]}, description: 'Report type'}, {in: query, name: startDate, schema: {type: string, format: date}, description: 'Report start date (defaults to 30 days ago)'}, {in: query, name: endDate, schema: {type: string, format: date}, description: 'Report end date (defaults to today)'}]
            responses: {'200': {description: 'Report generated successfully', content: {application/json: {schema: {type: object, properties: {success: {type: boolean, example: true}, reportType: {type: string, example: transactions}, period: {type: object, properties: {startDate: {type: string, format: date-time}, endDate: {type: string, format: date-time}}}, data: {type: array, description: 'Report data aggregated by the specified criteria'}}}}}}, '400': {description: 'Invalid report type'}, '403': {description: 'Admin access required'}}
...
    /api/reports:
        get:
            summary: 'Get user reports'
            tags: [Reports]
            security: [{bearerAuth: []}]
            parameters: [{in: query, name: page, schema: {type: integer, default: 1}, description: 'Page number'}, {in: query, name: limit, schema: {type: integer, default: 20}, description: 'Number of reports per page'}, {in: query, name: status, schema: {type: string, enum: [submitted, in_progress, resolved, closed]}, description: 'Filter by report status'}, {in: query, name: category, schema: {type: string}, description: 'Filter by report category'}]
            responses: {'200': {description: 'List of user reports', content: {application/json: {schema: {type: object, properties: {success: {type: boolean}, count: {type: integer}, total: {type: integer}, page: {type: integer}, pages: {type: integer}, data: {type: array, items: {$ref: '#/components/schemas/Report'}}}}}}}, '401': {description: Unauthorized}}
        post:
            summary: 'Submit a new report'
            tags: [Reports]
            security: [{bearerAuth: []}]
            requestBody: {required: true, content: {multipart/form-data: {schema: {type: object, required: [category, subject, description], properties: {category: {type: string, enum: [account_issue, transaction_problem, card_issue, loan_problem, technical_issue, complaint, suggestion, other]}, subject: {type: string, maxLength: 200}, description: {type: string, maxLength: 2000}, priority: {type: string, enum: [low, medium, high, urgent], default: medium}, images: {type: array, items: {type: string, format: binary}, maxItems: 3}}}}}}
            responses: {'201': {description: 'Report submitted successfully', content: {application/json: {schema: {type: object, properties: {success: {type: boolean}, message: {type: string}, data: {$ref: '#/components/schemas/Report'}}}}}}, '400': {description: 'Validation error'}, '401': {description: Unauthorized}}
    '/api/reports/{reportId}':
        get:
            summary: 'Get specific report by ID'
            tags: [Reports]
            security: [{bearerAuth: []}]
            parameters: [{in: path, name: reportId, required: true, schema: {type: string}, description: 'Report ID'}]
            responses: {'200': {description: 'Report details', content: {application/json: {schema: {type: object, properties: {success: {type: boolean}, data: {$ref: '#/components/schemas/Report'}}}}}}, '401': {description: Unauthorized}, '404': {description: 'Report not found'}}
        put:
            summary: 'Update report (add comments/update description)'
            tags: [Reports]
            security: [{bearerAuth: []}]
            parameters: [{in: path, name: reportId, required: true, schema: {type: string}, description: 'Report ID'}]
            requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {description: {type: string, maxLength: 2000}, priority: {type: string, enum: [low, medium, high, urgent]}}}}}}
            responses: {'200': {description: 'Report updated successfully'}, '401': {description: Unauthorized}, '404': {description: 'Report not found'}}
        delete:
            summary: 'Delete report'
            tags: [Reports]
            security: [{bearerAuth: []}]
            parameters: [{in: path, name: reportId, required: true, schema: {type: string}, description: 'Report ID'}]
            responses: {'200': {description: 'Report deleted successfully'}, '401': {description: Unauthorized}, '404': {description: 'Report not found'}}
...